fix(pagerduty): resolve worker CodeQL findings
