+++ src/dev_health_ops/providers/pagerduty/webhook_worker.py
@@ -1,21 +1,19 @@
 from __future__ import annotations
 
-from dataclasses import dataclass
 from uuid import UUID
 
-from dev_health_ops.credentials.types import PagerDutyCredentials
 from dev_health_ops.db import get_postgres_session
-from dev_health_ops.providers.pagerduty.auth import (
-    ApiTokenAuth,
-    OAuthBearerAuth,
-    PagerDutyAuth,
-)
 from dev_health_ops.providers.pagerduty.webhook_worker_graph import (
     _load_active_pagerduty_webhook_binding,
     hydrate_locked_pagerduty_webhook_auth,
     load_active_pagerduty_webhook_context,
     lock_active_pagerduty_webhook_graph,
     reconcile_pagerduty_webhook_with_locked_graph,
+)
+from dev_health_ops.providers.pagerduty.webhook_worker_shared import (
+    PagerDutyWebhookAuth,
+    PagerDutyWebhookWorkerContext,
+    build_pagerduty_webhook_auth,
 )
 
 __all__ = [
@@ -27,20 +25,6 @@ __all__ = [
     "reconcile_pagerduty_webhook_with_locked_graph",
     "resolve_pagerduty_webhook_binding",
 ]
-
-
-@dataclass(frozen=True, slots=True)
-class PagerDutyWebhookWorkerContext:
-    org_id: str
-    binding_id: str
-    provider_instance_id: str
-    credential_id: str
-
-
-@dataclass(frozen=True, slots=True)
-class PagerDutyWebhookAuth:
-    auth: PagerDutyAuth
-    region: str
 
 
 async def resolve_pagerduty_webhook_binding(
@@ -74,22 +58,4 @@ async def load_pagerduty_webhook_auth(
             return await hydrate_locked_pagerduty_webhook_auth(graph)
 
 
-def _pagerduty_webhook_auth(
-    credentials: PagerDutyCredentials,
-) -> PagerDutyWebhookAuth:
-    match credentials.auth_mode:
-        case "api_token":
-            api_token = credentials.api_token
-            if not api_token:
-                raise RuntimeError("pagerduty webhook credential has no API token")
-            auth: PagerDutyAuth = ApiTokenAuth(api_token)
-        case "oauth" | "client_credentials":
-            access_token = credentials.access_token
-            if not access_token:
-                raise RuntimeError("pagerduty webhook credential has no access token")
-            auth = OAuthBearerAuth(access_token)
-        case _:
-            raise RuntimeError(
-                "pagerduty webhook credential has an unsupported auth mode"
-            )
-    return PagerDutyWebhookAuth(auth=auth, region=credentials.region)
+_pagerduty_webhook_auth = build_pagerduty_webhook_auth
